separate usm_ndarray C-API tests
